+++ controle.financeiro/src/main/java/com/modelo/api/controle/financeiro/entity/Categoria.java
@@ -17,7 +17,7 @@ public class Categoria {
 
     @Id
     @GeneratedValue(strategy=GenerationType.IDENTITY)
-    private Long idCategoria;
+    private int idCategoria;
     private String nome;
 
     public Categoria() {
