+++ controle.financeiro/src/main/java/com/modelo/api/controle/financeiro/repository/SubCategoriaRepository.java
@@ -4,6 +4,6 @@ import com.modelo.api.controle.financeiro.entity.SubCategoria;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SubCategoriaRepository extends JpaRepository<SubCategoria, Long> {
+public interface SubCategoriaRepository extends JpaRepository<SubCategoria, Integer> {
     
 }
